requirements.txt
psycopg2==2.5.4
django-grappelli==2.6.2
djangorestframework==2.4.2
markdown==2.5.1
django-filter=0.8
django-rest-swagger=0.1.14
Sphinx==1.2.3
sphinx_rtd_theme==0.1.5
django-oauth-toolkit==0.7.2
hashids==1.0.1
django-model-utils==2.2
